feat: Factory is now able create contract and set start date. refactored the project structure and naming for better readability
