--- backend/src/lang/dialog-flow.factory.ts
@@ -1,6 +1,0 @@
-import { EmploymentContractService } from 'database/employmentContract/employmentContract.service';
-import { ParameterHandler } from './parameter.abstract';
-
-export abstract class AbstractFactory {
-    public abstract createParameterHandler(): ParameterHandler;
-}
