+++ backend/src/lang/factory/factory.abstract.ts
@@ -1,0 +1,6 @@
+import { EmploymentContractService } from './../../database/employmentContract/employmentContract.service';
+import { ParameterHandler } from './parameter/parameter.abstract';
+
+export abstract class AbstractFactory {
+    public abstract createParameterHandler(): ParameterHandler;
+}
